微博：增加系统回收站不直接删除

+++ wojilu.Core/Common/Microblogs/Service/MicroblogService.cs
@@ -1,4 +1,4 @@
-/*
+Ôªø/*
  * Copyright (c) 2010, www.wojilu.com. All rights reserved.
  */
 
@@ -21,6 +21,7 @@ using wojilu.Common.Tags;
 using wojilu.Members.Users.Interface;
 using wojilu.Members.Users.Service;
 using wojilu.Web.Utils;
+using wojilu.Common.AppBase;
 
 namespace wojilu.Common.Microblogs.Service {
 
@@ -39,13 +40,17 @@ namespace wojilu.Common.Microblogs.Service {
             followerService = new FollowerService();
         }
 
+        private static String showCondition() {
+            return " and SaveStatus=" + SaveStatus.Normal;
+        }
+
         public virtual int CountByUser( int userId ) {
-            return Microblog.count( "UserId=" + userId );
+            return Microblog.count( "UserId=" + userId + showCondition() );
         }
 
         public virtual List<Microblog> GetByUser( int count, int userId ) {
             if (count <= 0) count = 10;
-            return Microblog.find( "User.Id=" + userId ).list( count );
+            return Microblog.find( "User.Id=" + userId + showCondition() ).list( count );
         }
 
         public virtual List<IBinderValue> GetMyRecent( int count, int userId ) {
@@ -58,37 +63,26 @@ namespace wojilu.Common.Microblogs.Service {
         }
 
         public virtual Microblog GetFirst( int userId ) {
-            return Microblog.find( "User.Id=" + userId ).first();
+            return Microblog.find( "User.Id=" + userId + showCondition() ).first();
         }
 
         public virtual List<Microblog> GetCurrent( int count, int userId ) {
             if (count <= 0) count = 1;
-            return Microblog.find( "User.Id=" + userId ).list( count );
+            return Microblog.find( "User.Id=" + userId + showCondition() ).list( count );
         }
 
         public virtual DataPage<Microblog> GetPageList( int userId, int pageSize ) {
 
-            DataPage<Microblog> list = Microblog.findPage( "UserId=" + userId, pageSize );
+            DataPage<Microblog> list = Microblog.findPage( "UserId=" + userId + showCondition(), pageSize );
             return list;
         }
 
-        public virtual DataPage<Microblog> GetPageListAll( int pageSize ) {
-
-            DataPage<Microblog> list = Microblog.findPage( "", pageSize );
-            return list;
-        }
-
-        public virtual DataPage<Microblog> GetPicPageListAll( int pageSize ) {
-
-            DataPage<Microblog> list = Microblog.findPage( "Pic <>'' ", pageSize );
-            return list;
-        }
 
         public virtual DataPage<Microblog> GetFollowingPage( int ownerId, int pageSize ) {
 
             String followingIds = getFriendAndFollowingIds( ownerId );
 
-            return Microblog.findPage( "UserId in (" + followingIds + ")", pageSize );
+            return Microblog.findPage( "UserId in (" + followingIds + ")" + showCondition(), pageSize );
 
         }
 
@@ -97,8 +91,8 @@ namespace wojilu.Common.Microblogs.Service {
             String followingIds = getFriendAndFollowingIds( ownerId );
             searchKey = strUtil.SqlClean( searchKey, 10 );
             if( strUtil.IsNullOrEmpty( searchKey) )
-                return Microblog.findPage( "UserId in (" + followingIds + ")" );
-            return Microblog.findPage( "UserId in (" + followingIds + ") and Content like '%"+searchKey+"%'" );
+                return Microblog.findPage( "UserId in (" + followingIds + ")" + showCondition() );
+            return Microblog.findPage( "UserId in (" + followingIds + ") and Content like '%" + searchKey + "%'" + showCondition() );
 
         }
 
@@ -138,14 +132,14 @@ namespace wojilu.Common.Microblogs.Service {
             
             if( i==0 ) {
 
-                // ±£¥Êtag
+                // ‰øùÂ≠òtag
                 TagService.SaveDataTag( blog, mp.GetTagList() );
 
-                // ∑¢Õ®÷™
+                // ÂèëÈÄöÁü•
                 addNotification( smbinder.GetValidUsers(), blog );
             }
 
-            // ◊™∑¢–Ë“™À¢–¬‘≠Ã˚µƒ◊™∑¢¡ø
+            // ËΩ¨ÂèëÈúÄË¶ÅÂà∑Êñ∞ÂéüÂ∏ñÁöÑËΩ¨ÂèëÈáè
             if (blog.ParentId > 0) {
                 Microblog parent = GetById( blog.ParentId );
                 if (parent != null) {
@@ -168,11 +162,8 @@ namespace wojilu.Common.Microblogs.Service {
 
         private void addNotification( List<User> users, Microblog blog ) {
 
-            // ∏¯@”√ªß∑¢Õ®÷™
+            // Áªô@Áî®Êà∑ÂèëÈÄöÁü•
             foreach (User u in users) {
-
-                //String msg = string.Format( "”–Œ¢≤©Ã·µΩ¡Àƒ˙:<a href=\"{0}\">{1}</a>", lnk, strUtil.ParseHtml( blog.Content, 30 ) );
-                //nfService.send( u.Id, msg );
 
                 MicroblogAt mat = new MicroblogAt();
                 mat.Microblog = blog;
@@ -196,15 +187,15 @@ namespace wojilu.Common.Microblogs.Service {
 
             feed.Ip = log.Ip;
 
-            // ◊™∑¢Œ¢≤©–≈œ¢
+            // ËΩ¨ÂèëÂæÆÂçö‰ø°ÊÅØ
             String pbody = "";
             if (log.ParentId > 0) {
                 Microblog parent = GetById( log.ParentId );
                 if (parent == null) {
-                    pbody = " [±ª◊™Œ¢≤©“—±ª‘≠◊˜’ﬂ…æ≥˝]";
+                    pbody = " [Ë¢´ËΩ¨ÂæÆÂçöÂ∑≤Ë¢´Âéü‰ΩúËÄÖÂà†Èô§]";
                 }
                 else {
-                    pbody = ": [◊™]" + parent.Content;
+                    pbody = ": [ËΩ¨]" + parent.Content;
                 }
             }
 
@@ -222,9 +213,15 @@ namespace wojilu.Common.Microblogs.Service {
             feedService.publishUserAction( feed );
         }
 
+        //----------------------------------------------------------------------------
 
         public virtual void Delete( Microblog blog ) {
+
+            if (blog == null) throw new ArgumentNullException( "blog" );
+
+            blog.SaveStatus = SaveStatus.Delete;
             blog.delete();
+
         }
 
         public virtual void DeleteBatch( string ids ) {
@@ -232,7 +229,7 @@ namespace wojilu.Common.Microblogs.Service {
             int[] arrIds = cvt.ToIntArray( ids );
             if (arrIds.Length == 0) return;
 
-            Microblog.deleteBatch( "id in (" + ids + ")" );
+            Microblog.updateBatch( "SaveStatus=" + SaveStatus.Delete, "id in (" + ids + ")" );
         }
 
     }
